Add rogue exceptions & loggers and do some cleanup
