+++ src/rogue/protocols/xilinx/module.cpp
@@ -26,7 +26,6 @@
 #include <rogue/protocols/xilinx/module.h>
 #include <rogue/protocols/xilinx/Xvc.h>
 #include <rogue/protocols/xilinx/JtagDriver.h>
-#include <rogue/protocols/xilinx/JtagDriverAxisToJtag.h>
 
 namespace bp = boost::python;
 namespace rpx = rogue::protocols::xilinx;
@@ -44,6 +43,5 @@ void rpx::setup_module()
    bp::scope io_scope = module;
 
    rpx::JtagDriver::setup_python();
-   rpx::JtagDriverAxisToJtag::setup_python();
    rpx::Xvc::setup_python();
 }
